+++ main.js
@@ -312,6 +312,8 @@ function createMainWindow() {
     height,
     frame: false,
     transparent: true,
+    backgroundColor: "#00000000",
+    thickFrame: false,
     alwaysOnTop: true,
     resizable: false,
     hasShadow: false,